http_communication_protocols/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.4
six==1.17.0
sseclient-py==1.8.0
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.23.0
